fix rating up heroku

+++ database/migrations/2017_04_15_005027_edit_avg_rating_to_products.php
@@ -14,7 +14,7 @@ class EditAvgRatingToProducts extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('avg_rating')->nullable()->change();
+            $table->float('avg_rating')->nullable()->change();
         });
     }
